requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=9.0.0
omegaconf>=2.1.0
huggingface_hub>=0.12.0 